Move test to distributed query tests
This test relies on a layout being picked during planning. For
LocalQueryRunner this only happens because of a synthetic rule
(PickLayout w/o predicate) that attaches a layout to the table scan.
So, in a sense, it's just a coincidence that it works.

In distributed execution, the job is done by AddExchanges, so
we want to make sure we're testing that behavior.
